Add WritablePolyshape#addVertices method
This method allows to add a collection of vertices all at once.
One big advantage of supporting this is performance: it avoids repeated
reallocation of internal data structures, since it is known in advance
how many vertices are intended to be added in the batch.

For DefaultWritablePolygon2D, the x and y Trove-based fields needed
extension to support one-shot right-shifting of the arrays to enable
performant mass injection of new values.

For DefaultWritablePolyline, the vertices field needed more specific
typing on ArrayList to make use of its ensureCapacity method.