Improve GeomMasks javadoc
The javadoc was promising specific subtypes would be returned, but the
methods were not actually typed on those specific subtypes. Let's back
off on those promises, and instead explain things in an interface-driven
way: in general, the closed* methods give a ROI with CLOSED boundary
type, while the open* methods give a ROI with OPEN boundary type.

This commit does narrow the return type of kDTreeRealPointCollection and
realPointSampleListRealPointCollection, since it is helpful (especially
for the RealPointSampleListRealPointCollection) to have the properly
typed object for subsequent API calls like addPoint.