Fix ImgLabeling for various types of index images
The accessors for ImgLabeling are build with the assumption that
RandomAccess.get() and Cursor.get() of the index image will always return
the same object. This constraint holds for ArrayImg and CellImg, but fails
for other RandomAccessibleIntervals like:
* ListImg,
* ConvertedRandomAccessilbeInterval,
* StackView
This commit removes the assumption, and therefor fixes ImgLabeling for the
previously mentioned RandomAccessibleIntervals.